Reorganize frontend into modular ES module architecture
- Split monolithic main.css into 5 focused stylesheets (base, navbar, home, search, statistics)
- Decompose 469-line search.js into 4 ES modules (search, fields, results, viewer)
- Extract shared config, API wrapper, and navbar into reusable js/ modules
- Move static assets (favicon, logo, images) into dedicated assets/ directory
- Convert all pages from script-defer to ES module loading
- Remove duplicated navbar HTML in favor of dynamic JS-rendered navbar
- Update all HTML files with new CSS/asset paths
