+++ css/navbar.css
@@ -1,0 +1,41 @@
+.navbar {
+    width: 100%;
+    height: 3.5rem;
+    font-size: 1.1rem;
+    background-color: #dd0428;
+    box-shadow: 0 0.1rem 0.5rem #0006;
+    display: flex;
+    flex-flow: row nowrap;
+    position: fixed;
+    top: 0;
+    z-index: 1;
+}
+.navbar a {
+    height: 100%;
+    color: #fffd;
+    text-decoration: none;
+    user-select: none;
+    display: inline-flex;
+    align-items: center;
+    padding: 1.1rem 1.1rem 1rem;
+    box-sizing: border-box;
+    transition: color 0.15s, background-color 0.15s;
+}
+.navbar a:hover {
+    color: #ffff;
+    background-color: #e85971;
+}
+.navbar-logo {
+    padding: 0.6rem !important;
+    padding-right: 1rem !important;
+}
+.navbar-logo img {
+    filter: grayscale(1) contrast(10) invert(1);
+}
+.navbar-logo span {
+    padding-top: 0.1rem;
+    padding-left: 0.75rem;
+}
+.navbar-right {
+    margin-left: auto;
+}
