+++ css/search.css
@@ -1,188 +1,3 @@
-body {
-    font-family: Tahoma, sans-serif;
-    margin: 0;
-    padding: 0;
-}
-
-.navbar {
-    width: 100%;
-    height: 3.5rem;
-    font-size: 1.1rem;
-    background-color: #dd0428;
-    box-shadow: 0 0.1rem 0.5rem #0006;
-    display: flex;
-    flex-flow: row nowrap;
-    position: fixed;
-    top: 0;
-    z-index: 1;
-}
-.navbar a {
-    height: 100%;
-    color: #fffd;
-    text-decoration: none;
-    user-select: none;
-    display: inline-flex;
-    align-items: center;
-    padding: 1.1rem 1.1rem 1rem;
-    box-sizing: border-box;
-    transition: color 0.15s, background-color 0.15s;
-}
-.navbar a:hover {
-    color: #ffff;
-    background-color: #e85971;
-}
-.navbar-logo {
-    padding: 0.6rem !important;
-    padding-right: 1rem !important;
-}
-.navbar-logo img {
-    filter: grayscale(1) contrast(10) invert(1);
-}
-.navbar-logo span {
-    padding-top: 0.1rem;
-    padding-left: 0.75rem;
-}
-.navbar-right {
-    margin-left: auto;
-}
-
-.content {
-    font-family: Verdana, sans-serif;
-    margin-top: 3rem;
-    padding: 2rem;
-}
-.content a:link, .common-activate {
-    color: #045edd;
-    text-decoration: none;
-}
-.content a:hover, .common-activate:hover {
-    text-decoration: underline;
-    cursor: pointer;
-}
-.content a:active, .common-activate:active {
-    color: #dd0428;
-    text-decoration: underline;
-}
-.content a:visited {
-    color: #ed687e;
-    text-decoration: none;
-}
-.content input, .content button, .content select {
-    padding: 0.2rem 0.3rem;
-}
-.content input:not([type]), .content input[type="text"] {
-    width: 11rem;
-}
-.content input[type="checkbox"], .content input[type="radio"] {
-    vertical-align: top;
-}
-.content input[type="checkbox"] ~ label {
-    margin-right: 1rem;
-}
-
-.common-center {
-    max-width: 50rem;
-    margin: 0 auto;
-}
-.common-header {
-    font-family: Tahoma, sans-serif;
-    font-size: 2rem;
-    margin-bottom: 1rem;
-    padding-bottom: 0.4rem;
-    border-bottom: 0.15rem solid #0003;
-}
-.common-subheader {
-    font-family: Tahoma, sans-serif;
-    font-size: 1.7rem;
-    margin: 1rem 0 0.5rem;
-}
-.common-text {
-    margin-bottom: 0.5rem;
-}
-.common-text + .common-text {
-    margin-top: 1rem;
-}
-.common-check {
-    padding: 0.2rem 0;
-    display: block;
-}
-.common-loading {
-    font-family: Tahoma, sans-serif;
-    font-size: 1.3rem;
-    text-align: center;
-}
-.common-loading > div {
-    width: 0.8rem;
-    height: 0.8rem;
-    display: inline-block;
-    margin-right: 0.2rem;
-    margin-bottom: 0.2rem;
-    border: 0.15rem solid #000;
-    border-radius: 1rem;
-    vertical-align: bottom;
-    animation: spin 1s linear infinite;
-}
-.common-loading > div > div {
-    width: 0.6rem;
-    height: 0.6rem;
-    background-color: #fff;
-    margin-left: -0.4rem;
-    margin-bottom: 0.3rem;
-}
-.common-table {
-    width: 100%;
-    table-layout: fixed;
-    border-collapse: collapse;
-}
-.common-table :is(td, th) {
-    padding: 0.5rem;
-    border: 0.15rem solid #0002;
-    overflow-wrap: anywhere;
-}
-.common-table th {
-    background-color: #0001;
-}
-.common-table ul {
-    margin: 0;
-    padding-left: 1rem;
-}
-
-.home-header {
-    width: 42rem;
-    height: 12rem;
-    white-space: nowrap;
-    margin: 0 auto;
-    display: flex;
-    gap: 2rem;
-}
-.home-text {
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    gap: 0.3rem;
-}
-.home-title {
-    font-family: Tahoma, sans-serif;
-    font-size: 3rem;
-    display: flex;
-    align-items: flex-end;
-    flex-grow: 1.5;
-}
-.home-subtitle {
-    flex-grow: 1;
-}
-.home-search {
-    text-align: center;
-    margin-top: 1rem;
-}
-.home-search-input, .home-search-button {
-    padding: 0.3rem 0.5rem !important;
-}
-.home-search-input {
-    width: 20rem !important;
-}
-
 .search {
     width: 30rem;
     height: calc(100% - 3.5rem);
@@ -418,36 +233,4 @@ body {
 }
 .download-table :is(td, th):nth-child(3) {
     width: 100%;
-}
-
-.stats-display {
-    margin-top: 1rem;
-    display: flex;
-    justify-content: center;
-    column-gap: 2em;
-}
-.stats-total {
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    flex: 0 0 12rem;
-}
-.stats-total div:last-child {
-    font-family: Tahoma, sans-serif;
-    font-size: 2rem;
-}
-.stats-platforms, .stats-platforms :is(td, th) {
-    width: auto !important;
-}
-.stats-platforms {
-    margin: 1rem auto 0;
-}
-.stats-platforms :is(td, th) {
-    text-align: center;
-    padding: 0.5rem 1rem;
-}
-
-@keyframes spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
 }
