+++ home.js
@@ -1,4 +1,9 @@
-fetch('https://db-api.unstable.life/stats').then(r => r.json()).then(json => {
+import { renderNavbar } from './js/navbar.js';
+import { fetchJson } from './js/api.js';
+
+renderNavbar('home');
+
+fetchJson('/stats').then(json => {
     document.querySelector('.home-games').textContent = json.libraryTotals[0].count.toLocaleString();
     document.querySelector('.home-animations').textContent = json.libraryTotals[1].count.toLocaleString();
 });
