+++ index.html
@@ -2,24 +2,16 @@
 <html>
     <head>
         <title>Flashpoint Database</title>
-        <link rel="icon" href="favicon.ico">
-        <link rel="stylesheet" href="main.css">
-        <script defer src="home.js"></script>
+        <link rel="icon" href="assets/favicon.ico">
+        <link rel="stylesheet" href="css/base.css">
+        <link rel="stylesheet" href="css/navbar.css">
+        <link rel="stylesheet" href="css/home.css">
+        <script type="module" src="home.js"></script>
     </head>
     <body>
-        <div class="navbar">
-            <a href="./" class="navbar-logo">
-                <img src="logo.svg" height="100%">
-                <span>Flashpoint Database</span>
-            </a>
-            <a href="search">Search</a>
-            <a href="statistics">Statistics</a>
-            <a href="https://github.com/FlashpointProject/flashpoint-database" class="navbar-right">Source Code</a>
-            <a href="https://bluemaxima.org/flashpoint">Back to Home</a>
-        </div>
         <div class="content">
             <div class="home-header">
-                <img src="logo.svg" height="100%" class="home-logo">
+                <img src="assets/logo.svg" height="100%" class="home-logo">
                 <div class="home-text">
                     <div class="home-title">Flashpoint Database</div>
                     <div class="home-subtitle">Search <b class="home-games"></b> games and <b class="home-animations"></b> animations</div>
