+++ js/api.js
@@ -1,0 +1,6 @@
+import { API_BASE } from './config.js';
+
+export async function fetchJson(endpoint) {
+    const response = await fetch(`${API_BASE}${endpoint}`);
+    return response.json();
+}
