+++ js/config.js
@@ -1,0 +1,4 @@
+export const API_BASE = 'https://db-api.unstable.life';
+export const IMAGES_BASE = 'https://infinity.unstable.life/images';
+export const ICONS_BASE = 'https://flashpointproject.github.io/flashpoint-database-logos/';
+export const RESULTS_PER_PAGE = 100;
