+++ search/index.html
@@ -2,21 +2,13 @@
 <html>
     <head>
         <title>Search - Flashpoint Database</title>
-        <link rel="icon" href="../favicon.ico">
-        <link rel="stylesheet" href="../main.css">
-        <script defer src="search.js"></script>
+        <link rel="icon" href="../assets/favicon.ico">
+        <link rel="stylesheet" href="../css/base.css">
+        <link rel="stylesheet" href="../css/navbar.css">
+        <link rel="stylesheet" href="../css/search.css">
+        <script type="module" src="search.js"></script>
     </head>
     <body>
-        <div class="navbar">
-            <a href="../" class="navbar-logo">
-                <img src="../logo.svg" height="100%">
-                <span>Flashpoint Database</span>
-            </a>
-            <a href="../search">Search</a>
-            <a href="../statistics">Statistics</a>
-            <a href="https://github.com/FlashpointProject/flashpoint-database" class="navbar-right">Source Code</a>
-            <a href="https://bluemaxima.org/flashpoint">Back to Home</a>
-        </div>
         <div class="content">
             <div class="search">
                 <div class="common-header">Search</div>
@@ -54,7 +46,7 @@
                         <span class="results-current-page"></span> / <span class="results-max-pages"></span>
                         <button class="results-forward-page">&#62;</button>
                         <button class="results-last-page">&#62;&#62;</button>
-                        
+
                         <input type="text" placeholder="#" class="results-input-page">
                         <button class="results-go-to-page">Go</button>
                     </div>
@@ -69,7 +61,7 @@
                     <span class="results-current-page"></span> / <span class="results-max-pages"></span>
                     <button class="results-forward-page">&#62;</button>
                     <button class="results-last-page">&#62;&#62;</button>
-                    
+
                     <input type="text" placeholder="#" class="results-input-page">
                     <button class="results-go-to-page">Go</button>
                 </div>
@@ -78,7 +70,7 @@
                         <div class="viewer-back common-activate">&#60; Back to Results</div>
                         <a class="viewer-play" target="_blank">Play at 9o3o</a>
                     </div>
-                    
+
                     <div class="common-subheader">Images</div>
                     <div class="viewer-images">
                         <fieldset class="viewer-logo">
@@ -90,10 +82,10 @@
                             <a target="_blank"><img></a>
                         </fieldset>
                     </div>
-                    
+
                     <div class="common-subheader">Metadata</div>
                     <table class="viewer-metadata common-table"></table>
-                    
+
                     <div class="viewer-add-apps-header common-subheader">Additional Apps</div>
                     <div class="viewer-add-apps"></div>
                 </div>
