+++ statistics/index.html
@@ -2,21 +2,13 @@
 <html>
     <head>
         <title>Statistics - Flashpoint Database</title>
-        <link rel="icon" href="../favicon.ico">
-        <link rel="stylesheet" href="../main.css">
-        <script defer src="statistics.js"></script>
+        <link rel="icon" href="../assets/favicon.ico">
+        <link rel="stylesheet" href="../css/base.css">
+        <link rel="stylesheet" href="../css/navbar.css">
+        <link rel="stylesheet" href="../css/statistics.css">
+        <script type="module" src="statistics.js"></script>
     </head>
     <body>
-        <div class="navbar">
-            <a href="../" class="navbar-logo">
-                <img src="../logo.svg" height="100%">
-                <span>Flashpoint Database</span>
-            </a>
-            <a href="../search">Search</a>
-            <a href="../statistics">Statistics</a>
-            <a href="https://github.com/FlashpointProject/flashpoint-database" class="navbar-right">Source Code</a>
-            <a href="https://bluemaxima.org/flashpoint">Back to Home</a>
-        </div>
         <div class="content">
             <div class="common-center">
                 <div class="common-header">Statistics</div>
@@ -31,7 +23,7 @@
                         <div></div>
                     </div>
                 </div>
-                
+
                 <div class="common-subheader">Total Entries by Format</div>
                 <div class="stats-display">
                     <div class="stats-legacy stats-total">
@@ -43,7 +35,7 @@
                         <div></div>
                     </div>
                 </div>
-                
+
                 <div class="common-subheader">Total Entries by Platform</div>
                 <table class="stats-platforms common-table"></table>
             </div>
